Expose the newly added time attributes for runs.
Run attributes for create, start and finish times were added to the REST
interface. Now exposed through the Ruby API.

Also, the method to wait for a run to complete was changed. Previously
Run#start_and_wait, now just Run#wait. So the run must be started first. This
was done as it allows the start time to be queried as soon as the run has been
started as opposed to when it has already finished.