Update the Run#mkdir method to cope with nested directories.
Previously if you tried to create a directory within a directory like this:
 run.mkdir("test")
 run.mkdir("test/test")
the second call would fail as the server would refuse to create a directory
called "test/test".

The Run#mkdir function has been updated to split such paths at the slash so that
the server sees the request as that of making a directory called "test" within
the "test" directory rather than the above situation.