Add exceptions and handlers for the Run class.
Most of the exceptions raised by the server can just pass up through the Run
class, but it does need one of its own to raise: RunStateError. This is raised
when the user tries to something which the Run is not able to do in its current
state. An example of this would be to set an input port when it is already
running.