Stop t2-run-workflow from deleting runs automatically.
This is useful if the user would like to be able to download the results of a
run at a later time. Note that the run is still subject to the server's default
removal strategy - the results are not kept in perpetuity, but not deleted as
soon as the run finishes.

Runs can still be deleted on completion by setting the -D or --delete switch on
the command line.