Change the Run#get_output method to correctly return a list.
Taverna treats outputs with multiple values as lists (of lists (of lists, etc))
but Taverna Server stores these values in the filesystem as lists represented by
directories and the list items as files. Previously this code required that the
user navigate through this structure and retrieve the individual data items
separately, building up the list themselves. This patch changes the
Run#get_output method so that it does this and builds up and returns the lists
automatically.

The run_workflow script is also updated to reflect this change.