Move to LibXML for parsing server responses.
The gem was using REXML (the built in Ruby library) but LibXML is orders of
magnitude faster. XPath queries are pre-compiled so that if they are used
multiple times a significant speed up can be expected.

In most cases the speed up noticed will be minimal, but when connecting to a
server instance with a lot of running jobs, querying those jobs will be much
quicker.

The Rakefile/gemspec has been updated to reflect the fact that libxml-ruby is
now a runtime dependency.