Refactor the code that builds output port lists.
This now builds up lists in the correct order but needs to know if it is at the
top level to do so. Run#ls is renamed to Run#_ls_ports to reflect the fact that
it is now only used for navigating into the output ports.

Test suite updated to reflect changes.