Support version 2 of the Server API.
This is a much more user orientated model. Most of the heavy lifting for v2
changes were done in the previous commit, "Abstract connection and credential
logic out of the Server class".

Remove @run_limit and make it requested each time - this is because different
users will see different numbers.

Check for version of remote server and compile links sets as appropriate.