Fix input value escaping to allow XML in inputs.
Input values were not being escaped so when passing XML in as an input value it
was not getting passed up to server correctly.

This is fixed by building the input message via the XML library.

An XML upload test is also added to the test suite to catch such things in the
future.