Make the baclava input/output calls more "ruby-like"
The previous versions of these calls took the C++/Java get and set form but this
is not particularly ruby-like so they have been deprecated. Their replacements
also more accurately reflect the fact that baclava output simply needs to be
turned on (or requested) and then the resultant file needs to be fetched when
the run is complete.

In addition, the t2-run-workflow script now requires a baclava output filename.