Abstract XML functions out of Server and Run classes.
The XML code was starting to get messy and spread out over the two main classes.
Moving it out into its own module and mixing in where required neatens it up and
also makes it easy to change the XML implementation in the future.

The XPath expression constants were also moved the other way - from the XML
module to the classes in which they are actually required. This was to stop the
namespace getting polluted unnecessarily and avoid clashes.

Additionally, exception handling in _ls_ports has been removed as I could not
work out why it was needed - testing did not raise any exceptions so maybe it is
not needed anymore.