Move the code to set a run's input into the Run class.
This logic was out of place in the Server class.

As a side effect Server#set_run_attribute changes slightly. It now requires that
a type for the value to be set is specified as opposed to assuming "text/plain".

Also, Run#inputs is no longer needed. It was a hack anyway. Deprecated.