Remove methods that access run attributes from the Server API.
Any method that just operates on an individual runs' internal attributes, such
as creating a directory or uploading a file is removed from the Server API. This
is because such methods are only provided in Server to allow for a single point
of entry to a shared connection. All functionality is provided by methods in the
Run API.

Also, Server#set_run_input and Server#set_run_input_file are deprecated.