Add zip output facilities to the Run class.
This simply exposes the server function by which the entire working directory of
a run can be downloaded as a single binary blob in zip format.

A simple test is also included. It doesn't test the correctness of the file,
just that it is non-empty and returned without errors.

Facilities are also added to access this functionality from the t2-run-workflow
script.