Update unit tests (inc. for version 1/2 differences).
Versions 1 and 2 of the server are now different enough that the tests are
starting to diverge further and further. Until we can drop support for version 1
it makes sense to keep the tests fully separate.

Other updates here include new workflows to test error handling in the new
OutputPort class and a file of text input to help test partial output value
downloading.