Add Run attribute creation and deletion facilities to Server.
Server#create_run_attribute attempts to create the requested resource on the Run
via POST and returns the resultant resource address as a string. Just the leaf
of the new resource name is returned. The Connection#POST method has been
refactored to allow for this new functionality.

Server#delete_run_attribute deletes the requested resource if it exists.