Facilities for granting and revoking permissions on runs.
This allows the owner of a run to grant other users permissions to read, update
and destroy the run. These permissions can also be revoked.

Updates were also needed to the server class to ensure that the caching of run
objects doesn't allow non-owners of a run to alter its security settings.