Enable dynamic selection of XML implementation.
Each implementation is tried in turn. If it can't be used because the required
library is not installed it tries the next one. Rexml is tried last and will
always work as rexml is part of the standard library. LibXML2 is tried first,
then Nokogiri, then rexml.

The Rakefile is also updated to remove libxml-ruby as a required runtime
dependency. It is now a development dependency along with nokogiri.