Update the baclava file input API.
This was slightly confusing due to the fact that inputs are generally set
individually unless baclava is used, in which case all of the inputs are set by
the same baclava file.

The setting of the baclava file has been made more Ruby-like and knowledge of
baclava has been removed from the individual input port objects.