Update the input port API.
Remove Run#set_input, Run#set_input_file and Run#upload_input_file - this was
not a particularly Ruby way of doing things and the difference between using a
remote file versus a local file was never really clear.

Setting of inputs is completely delegated to the InputPort class itself using
InputPort#value=, InputPort#remote_file= and InputPort#file= for above methods
respectively.

The old methods are marked as deprecated and will be removed in the next major
version.