Remove deprecated methods.
Now that the next version is 1.0.0 we need to remove all the methods and
functionality that we have deprecated. All such functionality was in the Run and
Server classes. An old top level require file with an inconsistent name is also
removed.

The t2-delete-runs script is updated as a result of this change.