Add methods to test for empty ports or port values.
OutputPort#empty? and PortValue#empty? added. This allows testing before
attempting to download or stream data from a port.

Note that if the output port holds a list then it is not considered empty, even
if that list is empty. This is because the port itself is not empty, there is a
list there! A separate test should be performed to see if that list is empty or
not.

Tests also added.