Add output port data streaming functionality.
Output port data can now be streamed from the server to allow for large data to
be handled without it all being downloaded into memory first. Any data that is
already in memory will not be downloaded again but is inserted into the stream
in the correct place.

Tests are also included.