Move interaction code out of Server, into Run.
This is because the interaction feeds are per-run now, rather than a single
global feed per server.

The Feed and Reader classes are merged now that each run has a feed and the
actual reading of the feed is done via the Run object's connection to the remote
server due to security requirements.