Fix well-hidden HTTPS connection setup bug.
Passing in no connection parameters was failing because, although the params
were being set up correctly in the superclass constructor, they weren't then
being used properly.

This was introduced when the SSL version parameter was added. It is difficult to
test for due to the fact that connections are cached once set up and tests are
run in random order.